replace fmt.Errorf with Exported Err type
Creates a new Err type that is able to represent the domain of the
error, as well as maintain a reference to the underlying error for
user reference.

add test to ensure that the error formats in a similar way to the
previous errors being returned.

add test to ensure that the new Err type is being returned from
Client.Endpoint when a failure occurs instead of an anonymous
errors.New error.