metrics: Move the Provider interface to metrics
Reasons:

1. The Provider interface is all about returning things from metrics. It makes
more sense to me to have it located there.

2. provider.Provider is a form of stutter. It's more accepted then other forms
though (cf. hash.Hash), but I personally dislike it. For instance, what is a
provider.Provider providing? metrics.Provider says a lot more IMO.

3. Importing the provider package incurrs a lot of additional dependencies, most
of which aren't needed. Our work around to this is to redefine a Provider
interface locally. These are usually defined in a local `metrics` package
requiring either the local or go-kit metrics package to be renamed during import
which is :-(. These local metrics packages proliferate over time across projects.