update to generic counter
based on issue 525, worked
on taking a stab at reworking the implementation of generic using space
from internal package. before getting too far into that, decided to see
if this solution was a good one by just working on the generic counter.
if it is, will update the other Metrics in a similar way. tests have
been added that address the exact problem raised in the issue.