Move the ClientAfter calls to before the decode (#1008)
* Move the ClientAfter calls to before the decode

ClientAfter is documented as running _prior_ to being decoded. But, this
was only partly true. They were run prior to decode, but after the
jsonrpc unmarshalling,

As the callback uses the bare response object, and at least for me, is
used to debug what's seen on the wire, this seems incorrect.

* tests