test(transport/nats): remove nats server address log line
The running bit is set way earlier than the listener
that s.Addr returns. This lead to panics on several
occasions when the listener was not yet set.

Moving after ReadyForConnections could be a solution,
but logging the nats server address does not add
much to the test anyway: the address is logged
if there is an error.

Signed-off-by: Mark Sagi-Kazar <mark.sagikazar@gmail.com>